a columns selection feature has been added. The search has been modified to search for relationships. The [flag, avatar] columns have been replaced with photo. The pages for tour guides, and their types, and their ratings or reviews have been finalized.

+++ app/Http/Controllers/Dashboard/CityController.php
@@ -31,17 +31,20 @@ class CityController extends Controller
 
         if ($created) {
             if ($request->has('save_and_add')) {
-                return redirect()->route('cities.create')->with('success', __('main.messages.city_created'));
+                return redirect()->route('cities.create')->with('success', __('main.messages.type_created', ['type' => __('main.city')]));
             }
-            return redirect()->route('cities.index')->with('success', __('main.messages.city_created'));
+            return redirect()->route('cities.index')->with('success', __('main.messages.type_created', ['type' => __('main.city')]));
         }
 
-        return redirect()->route('cities.index')->with('error', __('main.messages.city_creation_failed'));
+        return redirect()->route('cities.index')->with('error', __('main.messages.type_creation_failed', ['type' => __('main.city')]));
     }
 
     public function edit($id)
     {
-        $city = City::with(['country', 'state'])->findOrFail($id);
+        $city = City::with(['country', 'state'])->find($id);
+        if (!$city) {
+            return redirect()->back()->with('error', __('main.messages.not_found_this_type', ['type' => __('main.city')]));
+        }
         $states = State::orderBy('name')->get();
         $countries = Country::orderBy('name')->get();
         return view('pages.dashboard.cities.edit', compact('city', 'countries', 'states'));
@@ -49,25 +52,31 @@ class CityController extends Controller
 
     public function update(UpdateCitiesRequest $request, $id)
     {
-        $city = City::findOrFail($id);
+        $city = City::find($id);
+        if (!$city) {
+            return redirect()->back()->with('error', __('main.messages.not_found_this_type', ['type' => __('main.city')]));
+        }
         $validated = $request->validated();
 
         $updated = $city->update($validated);
         if ($updated) {
-            return redirect()->route('cities.index')->with('success', __('main.messages.city_updated'));
+            return redirect()->route('cities.index')->with('success', __('main.messages.type_updated', ['type' => __('main.city')]));
         }
 
-        return redirect()->route('cities.index')->with('error', __('main.messages.city_update_failed'));
+        return redirect()->route('cities.index')->with('error', __('main.messages.type_update_failed', ['type' => __('main.city')]));
     }
 
     public function destroy($id)
     {
-        $city = City::findOrFail($id);
+        $city = City::find($id);
+        if (!$city) {
+            return redirect()->back()->with('error', __('main.messages.not_found_this_type', ['type' => __('main.city')]));
+        }
         $deleted = $city->delete();
         if ($deleted) {
-            return redirect()->route('cities.index')->with('success', __('main.messages.city_deleted'));
+            return redirect()->route('cities.index')->with('success', __('main.messages.type_deleted', ['type' => __('main.city')]));
         }
 
-        return redirect()->route('cities.index')->with('error', __('main.messages.city_deletion_failed'));
+        return redirect()->route('cities.index')->with('error', __('main.messages.type_deletion_failed', ['type' => __('main.city')]));
     }
 }
